+++ 01/Task01.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import Timer from './../src/components/Timer';
+import Timer from '../src/containers/TimerContainer';
 
 const Task01 = () => (
     <section>
         <h2>Task 01</h2>
-        <Timer/>
+        <Timer />
     </section>
 );
